+++ .circleci/ci/it-tests.js
@@ -71,7 +71,7 @@ try {
 
     if (FT === 'true') {
         // add feature toggle impl bundle to check FT on cloud ready or release/650 instance
-        extras += ` --install-file ${buildPath}/it/core/src/main/resources/com.adobe.granite.toggle.impl.dev-1.2.0.jar`;
+        extras += ` --install-file ${buildPath}/it/core/src/main/resources/com.adobe.granite.toggle.impl.dev-1.1.2.jar`;
     }
 
     // Set an environment variable indicating test was executed
