+++ it/config/src/main/content/jcr_root/apps/system/config/com.adobe.granite.toggle.impl.dev.DynamicToggleProviderImpl.cfg.json
@@ -22,7 +22,6 @@
     "FT_FORMS-14255",
     "FT_FORMS-14068",
     "FT_FORMS-16351",
-    "FT_FORMS-14518",
-    "FT_FORMS-13519"
+    "FT_FORMS-14518"
   ]
 }
