--- it/content/src/main/content/jcr_root/conf/forms/core-components-it/samples/customfunctionedgedelivery/settings/cloudconfigs/.content.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:sling="http://sling.apache.org/jcr/sling/1.0"
-    jcr:primaryType="sling:Folder"/>
