+++ ui.af.apps/src/main/content/jcr_root/apps/core/fd/components/form/wizard/v1/wizard/clientlibs/commons/js/common.js
@@ -150,9 +150,6 @@
              * @param {Number} index The index of the tab to navigate to
              */
             navigate(index) {
-                if (this._active === index) {
-                    return; // already on this tab
-                }
                 this._active = index;
                 this.addSteppedClass();
                 this.refreshActive();
