+++ ui.tests/test-module/specs/tabsontop/tabsontop.runtime.cy.js
@@ -167,22 +167,6 @@ describe("Form with Tabsontop Layout Container with focus", () => {
 
     });
 
-    it("focus should be on the first component when navigated to the next tab", () => {
-        const [id, fieldView] = Object.entries(formContainer._fields)[0];
-        tab2().click().then(() => {
-            tab2().should('have.class', 'cmp-tabs__tab--active');
-            tab2().should('have.attr', 'aria-selected', 'true');
-            cy.get('input[name="textinputft2"]').should('be.focused');
-            tab1().should('have.attr', 'aria-selected', 'false');
-        });
-        tab1().click().then(() => {
-            tab1().should('have.class', 'cmp-tabs__tab--active');
-            tab1().should('have.attr', 'aria-selected', 'true');
-            cy.get('input[name="textinputft1"]').should('be.focused');
-            tab2().should('have.attr', 'aria-selected', 'false');
-        });
-    })
-
 
 });
 describe("Form with TabsOnTop Layout Container with Hidden Children", () => {
