+++ ui.tests/test-module/specs/tabsontop/tabsontop.runtime.repeatability.cy.js
@@ -181,58 +181,4 @@ describe("Form with TabsOnTop Container", () => {
     })
   })
 
-   it("Focus should be on field when it is repeated", () => {
-     getTabs().should('have.length', 4);
-     getTabPanels().should('have.length', 4);
-     cy.get("button").contains("+R1").click().then(() => {
-      getTabs().should('have.length', 5);
-      getTabPanels().should('have.length', 5);
-      getTabAtIndex(1).should('have.class', 'cmp-tabs__tab--active');
-      getTabPanelAtIndex(1).should('have.class', 'cmp-tabs__tabpanel--active');
-      getTabPanelAtIndex(1).find('input').should('be.focused');  
-   })
-     cy.get("button").contains("+R2").click().then(() => {
-      getTabs().should('have.length', 6);
-      getTabPanels().should('have.length', 6);
-      getTabAtIndex(4).should('have.class', 'cmp-tabs__tab--active');
-      getTabPanelAtIndex(4).should('have.class', 'cmp-tabs__tabpanel--active');
-      getTabPanelAtIndex(4).find('input').should('be.focused');  
-   })
-    cy.get("button").contains("+R3").click().then(() => {
-     getTabs().should('have.length', 7);
-     getTabPanels().should('have.length', 7);
-     getTabAtIndex(6).should('have.class', 'cmp-tabs__tab--active');
-     getTabPanelAtIndex(6).should('have.class', 'cmp-tabs__tabpanel--active');
-     getTabPanelAtIndex(6).find('select.cmp-adaptiveform-dropdown__widget').should('be.focused');  
-   })
-   cy.get("button").contains("-R3").click().then(() => {
-    getTabs().should('have.length', 6);
-    getTabPanels().should('have.length', 6);
-    getTabAtIndex(5).should('have.class', 'cmp-tabs__tab--active');
-    getTabPanelAtIndex(5).should('have.class', 'cmp-tabs__tabpanel--active');
-    getTabPanelAtIndex(5).find('select.cmp-adaptiveform-dropdown__widget').should('be.focused');  
-  })
-  cy.get("button").contains("-R2").click().then(() => {
-    getTabs().should('have.length', 5);
-    getTabPanels().should('have.length', 5);
-    getTabAtIndex(3).should('have.class', 'cmp-tabs__tab--active');
-    getTabPanelAtIndex(3).should('have.class', 'cmp-tabs__tabpanel--active');
-    getTabPanelAtIndex(3).find('input').should('be.focused');  
-  })
-  cy.get("button").contains("-R1").click().then(() => {
-    getTabs().should('have.length', 4);
-    getTabPanels().should('have.length', 4);
-    getTabAtIndex(0).should('have.class', 'cmp-tabs__tab--active');
-    getTabPanelAtIndex(0).should('have.class', 'cmp-tabs__tabpanel--active');
-    getTabPanelAtIndex(0).find('input').should('be.focused');  
-  })
-  cy.get("button").contains("-R1").click().then(() => {
-    getTabs().should('have.length', 3);
-    getTabPanels().should('have.length', 3);
-    getTabAtIndex(0).should('have.class', 'cmp-tabs__tab--active');
-    getTabPanelAtIndex(0).should('have.class', 'cmp-tabs__tabpanel--active');
-    getTabPanelAtIndex(0).find('input').should('be.focused');  
-  })
- })
-
 })
