+++ ui.tests/test-module/specs/wizard/wizard.runtime.repeatability.cy.js
@@ -278,8 +278,8 @@ describe('visibility of navigation buttons', function () {
                 // check if repeatable instance is removed honouring rule
                 // first tab gets active and previous nav button is invisible
                 // and next nav button is visible
-                getWizardPanelAtIndex(1).should('have.class', wizardPanelActive);
-                cy.get(previousNavButton).should('have.attr', 'data-cmp-visible', 'true');
+                getWizardPanelAtIndex(0).should('have.class', wizardPanelActive);
+                cy.get(previousNavButton).should('have.attr', 'data-cmp-visible', 'false');
                 cy.get(nextNavButton).should('have.attr', 'data-cmp-visible', 'true');
             });
         });
